requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
python-multipart==0.0.6

# Pydantic & Validation
pydantic==2.5.0

# Google Generative AI (Gemini)
google-generativeai==0.3.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# PDF Processing
pypdf==4.0.1

# Image Processing (Optional for OCR)
Pillow==10.1.0
pytesseract==0.3.10

# HTTP & Network
httpx==0.25.1
requests==2.31.0
urllib3==2.1.0

# Async Support
anyio==3.7.1
sniffio==1.3.0

# Data Serialization
python-json-logger==2.0.7

# Environment Variables
python-dotenv==1.0.0

# Utility Libraries
typing-extensions==4.8.0
packaging==23.2

# Security
cryptography==41.0.7

# Development (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
mypy==1.7.1